requirements.txt
datasets
pandas 
requests 
tqdm 
opencv-python 
albumentations 
pillow
imagehash
numpy
torch
torchvision
streamlit>=1.28.0
faiss-cpu>=1.7.4
